Split classes to different files
